Broken down AboutOperators to separate Arithmetic and Bitwise Operator Koans. Cleaned up conditionals.
